+++ CHANGELOG.md
@@ -5,13 +5,13 @@
 ### Patch Changes
 
 - 775f1fd: Add auto refresh to `userinfo`, `introspect`, `resource`, `status` and `claims` endpoints
-- 775f1fd: (Breaking) Changed `getEndpoints()` and `getAuthHook()` to getter methods `endpoints` and `authHook`
-- 775f1fd: (Breaking) Move hook settings to client options
+- 775f1fd: **[BREAKING CHANGE]** Changed `getEndpoints()` and `getAuthHook()` to getter methods `endpoints` and `authHook`
+- 775f1fd: **[BREAKING CHANGE]** Move hook settings to client options
 
 ## 0.1.9
 
 ### Patch Changes
 
 - 05731b2: Add defineConfig
-- 05731b2: (Breaking) Change factory method name from `create` to `factory`
+- 05731b2: **[BREAKING CHANGE]** Change factory method name from `create` to `factory`
 - 6797fa2: Update docs